requirements.txt
langchain
langchain[google-genai]
langchain-google-genai
langchain-community
langchain-text-splitters
langgraph
sqlalchemy
psycopg2-binary
pgvector
google-generativeai
pymupdf
pypdf
python-dotenv